% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tractor-class.R
\name{tractor}
\alias{tractor}
\title{Create a TRACTOR object}
\usage{
tractor(
  table,
  ranks = NULL,
  attributes = NULL,
  clustering = NULL,
  ties = list(),
  ordinations = list(),
  richness = NULL
)
}
\arguments{
\item{table}{matrix/data.frame of raw counts (taxa × samples)}

\item{ranks}{data.frame of taxonomic ranks (rows = taxa)}

\item{attributes}{data.frame of sample metadata (rows = samples)}

\item{clustering}{list or factor of cluster memberships / dendrograms}

\item{ties}{named list of dist objects / distance matrices}

\item{ordinations}{named list of ordination objects (stats / vegan)}

\item{richness}{data.frame of α-diversity metrics per sample}
}
\value{
An object of class \if{html}{\out{<tractor>}}
}
\description{
Create a TRACTOR object
}
