% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ties-helpers.R
\name{tractor_dist}
\alias{tractor_dist}
\title{Compute and store a distance matrix inside a tractor object}
\usage{
tractor_dist(x, method = "bray", name = method, ...)
}
\description{
Compute and store a distance matrix inside a tractor object
}
