% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination-helpers.R
\name{tractor_ordinate}
\alias{tractor_ordinate}
\title{Run PCoA or NMDS and attach to object}
\usage{
tractor_ordinate(x, tie = "bray", method = c("PCoA", "NMDS"), ...)
}
\description{
Run PCoA or NMDS and attach to object
}
